Create Program4.c
This program calculates the average of a set of numbers entered by the user.
It first asks the user how many numbers they want to enter. Then it dynamically allocates memory to store those numbers using malloc(). After that, the user enters each element, which is stored in the array.

A separate function named Average() receives this array and its size. Inside this function, all elements are added together using a loop. The total sum is then divided by the number of elements to get the average. The result is returned to the main function as a floating-point value.

Finally, the program displays the computed average and frees the allocated memory.