Create Program4.c
This program accepts N integers from the user and displays all elements that are divisible by both 3 and 5. First, the user is asked to enter how many numbers they want to input. Memory is then allocated dynamically to store these numbers. After the user enters all the elements, the Display function is called. This function iterates through the array and checks each number using the condition (Arr[iCnt] % 3 == 0) && (Arr[iCnt] % 5 == 0), which ensures the number is divisible by both 3 and 5. All such numbers are printed on the screen. Finally, the program frees the allocated memory before ending.