Create Program5.c
This program accepts N integers from the user and displays all the numbers that are multiples of 11. First, the program asks the user how many elements they want to enter and allocates memory dynamically to store those numbers. After the user enters all the values, the Display function is called. This function checks each number using the condition (Arr[iCnt] % 11 == 0) to determine whether it is divisible by 11 without any remainder. If the number satisfies the condition, it is printed on the screen. Finally, the program releases the allocated memory before ending.