Create Program1.c
This program prints a sequence of capital alphabets based on the number provided by the user. The user first enters a value representing how many characters should be displayed. Inside the Pattern function, a loop runs from 1 to the given number, starting with the character 'A'. During each iteration, the program prints the current character and then increments it to the next alphabet using ch++. As a result, if the user enters 5, the program prints: A B C D E. After finishing the loop, it prints a newline and the program ends.